backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
python-multipart
python-jose[cryptography]
passlib[bcrypt]
requests
pandas
google-generativeai
langchain
langchain-google-genai
chromadb